refactor(reposicao): split AdminReposicaoCadeiaLogistica (677→115)
Extrai a lógica (3 queries, 4 mutations, ltTotalAtualForn/recalcularComImpacto,
etapasPorForn, toggleExp) para o hook useCadeiaLogistica e quebra o JSX em
FornecedorCadeiaCard (card expansível + tabela de etapas) e HistoricoCard.
Comportamento preservado 1:1; +10 testes.
